src/dbControlByFlask/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
gevent==21.12.0
gunicorn==20.1.0
importlib-metadata==5.0.0
install-jdk==0.3.0
itsdangerous==2.1.2
Flask-SQLAlchemy==2.5.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
mysql-connector-python==8.0.30
mysql-connector-python-rf==2.2.2
pandas==1.5.0
pymysql==1.0.2
SQLAlchemy==1.4.17
tabula==1.0.5
tabula-py==2.5.1
tabulate==0.8.10
Werkzeug==2.2.2
zipp==3.8.1
